1st snackbar refactor: extracting components
