+++ src/components/SnackbarAlerts.jsx
@@ -1,0 +1,28 @@
+import {forwardRef} from 'react'
+
+import MuiAlert from '@mui/material/Alert';
+import Snackbar from '@mui/material/Snackbar';
+
+const Alert = forwardRef(function Alert(props, ref) {
+    return <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />;
+});
+
+function SnackbarAlerts({snackOpen, setSnackOpen, snackSeverity, snackMessage}) {
+
+    const handleSnackClose = (event, reason) => {
+        if (reason === 'clickaway') {
+            return;
+        }
+        setSnackOpen(false);
+    };
+
+    return (
+        <Snackbar open={snackOpen} autoHideDuration={6000} onClose={handleSnackClose} anchorOrigin={{ vertical: 'top', horizontal: 'center' }}>
+          <Alert onClose={handleSnackClose} severity={snackSeverity} sx={{ width: '100%' }}>
+            {snackMessage}
+          </Alert>
+        </Snackbar>
+    )
+}
+
+export default SnackbarAlerts
